Created via Snowpark/requirements.txt
cryptography
snowflake-snowpark-python[pandas]
snowflake-ingest
streamlit
pandas
regex
xlrd
faker
pmdarima
scikit-learn
